[Housing] WIP Furniture
Enabled the functionality of the allow furniture recover flag of your house. (this doesn't yet allow all to recover all items)
Adjusted how furniture doodad is "attached" to the house.
Added seller and sell price to do dirty flagging of the house.
Changed /pos to show details on any selected object or doodad (by providing the object id of it)